pyproject.toml
[project]
name = "feijoabot"
version = "1.0.0"
description = "Generalized miscellaneous discord bot"
readme = "README.md"
requires-python = ">=3.14"

dependencies = [
  "python-dotenv==1.1.1",
  "discord-py==2.6.3",
  "aiosqlite==0.21.0",
  "blackjack21>=4.0.0",
  "requests>=2.32.5",
]

[tool.ruff]
line-length = 130
indent-width = 4

[tool.ruff.lint]
# Require explicit selection of preview rules.
explicit-preview-rules = true
select = ["ALL"]
extend-select = ["ALL"]
extend-safe-fixes = ["ALL"]
extend-ignore = [
  "C901",
  "T201",    # Ignore print
  "DTZ005",
  "N806",    # allow dunder
  "S311",    # allow non-crypto random
  "INP001",  # implicit namespace
  "D100",    # Missing Docstring
  "D101",    # Missing Docstring
  "D102",    # Missing Docstring
  "D103",    # Missing Docstring
  "D107",    # Missing Docstring
  "FBT001",  # boolean-default-value-positional-argument
  "FBT002",  # boolean-default-value-positional-argument
  "FBT003",  # boolean-default-value-positional-argument
  "ERA001",  # Commented out code
  "PLR2004", # Magic numbers
  "PLR0913", # Too many arguments
  "TRY301",  # Abstract `raise` to an inner function
  "SLF001",  # Allow accessing private member (blackjack21 support)
]
preview = true

[tool.ty.rules]
unresolved-import = "ignore"

[tool.uv.sources]
discord-py = { git = "https://github.com/Rapptz/discord.py" }

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
